Add work

+++ starter_kit/exact_diag.py
@@ -2,15 +2,15 @@ from __future__ import annotations
 
 from functools import lru_cache
 
-import numpy as np
+import jax.numpy as np
 import pennylane as qml
 
 from .annni import build_annni_hamiltonian
 from .observables import order_parameter_summary
 
-I2 = np.eye(2, dtype=np.complex128)
-XMAT = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.complex128)
-ZMAT = np.array([[1.0, 0.0], [0.0, -1.0]], dtype=np.complex128)
+I2 = np.eye(2, dtype=np.complex64)
+XMAT = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.complex64)
+ZMAT = np.array([[1.0, 0.0], [0.0, -1.0]], dtype=np.complex64)
 
 
 def _kron_all(factors: list[np.ndarray]) -> np.ndarray:
@@ -36,11 +36,13 @@ def _pair_zz_matrix(n_qubits: int, site_a: int, site_b: int) -> np.ndarray:
     return _kron_all(factors)
 
 
-def _dense_annni_matrix(n_qubits: int, kappa: float, h: float, periodic: bool) -> np.ndarray:
+def _dense_annni_matrix(
+    n_qubits: int, kappa: float, h: float, periodic: bool
+) -> np.ndarray:
     limit_nn = n_qubits if periodic else n_qubits - 1
     limit_nnn = n_qubits if periodic else n_qubits - 2
     dimension = 2**n_qubits
-    matrix = np.zeros((dimension, dimension), dtype=np.complex128)
+    matrix = np.zeros((dimension, dimension), dtype=np.complex64)
 
     for i in range(limit_nn):
         matrix -= _pair_zz_matrix(n_qubits, i, (i + 1) % n_qubits)
@@ -58,8 +60,12 @@ def exact_ground_state(
     periodic: bool = True,
 ) -> dict[str, object]:
     """Return the lowest-energy eigenpair and observable summary."""
-    hamiltonian = build_annni_hamiltonian(n_qubits=n_qubits, kappa=kappa, h=h, periodic=periodic)
-    matrix = _dense_annni_matrix(n_qubits=n_qubits, kappa=kappa, h=h, periodic=periodic)
+    hamiltonian = build_annni_hamiltonian(
+        n_qubits=n_qubits, kappa=kappa, h=h, periodic=periodic
+    )
+    matrix = _dense_annni_matrix(
+        n_qubits=n_qubits, kappa=kappa, h=h, periodic=periodic
+    ).astype(np.float32)
     eigenvalues, eigenvectors = np.linalg.eigh(matrix)
     ground_state = eigenvectors[:, 0]
     summary = order_parameter_summary(ground_state, n_qubits)
